修改基础工具

+++ src/main/java/com/tinklabs/handy/base/vo/BaseVO.java
@@ -18,14 +18,13 @@ public class BaseVO implements Serializable {
 	/**
 	 * 主键
 	 */
-	private Integer id;
+	private Long id;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
-	
 }
